==> Industrial Landuse x Creative Economy/Scripts/1_Industrial.Parcels.Cleaning.Rmd <==
---
title: "1_Industrial.Parcels.Cleaning"
author: "Brandon Stanaway"
date: "2024-03-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

require(tidyverse)
require(tidycensus)
require(lubridate)
require(stringr)
require(janitor)
require(sf)
require(mapcdatakeys)
require(dplyr)
require(readxl)

#
keys <- mapcdatakeys::all_muni_data_keys |> 
  select(
    c(
      muni_name,
      muni_id,
      mapc,
      subrg_acr,
      cmtyp08,
      cmsbt08
    )
  ) |> 
  mutate(
    muni_name = stringr::str_to_title(muni_name)
  )

#Global options to remove scientific notation.
options(scipen = 999)
set.seed(351)

#Set root variables for local data ingestion
parcel_root <- "C:/Project_Work/Local_Data/Parcel_DB/"

```

Land Parcel Database 2018
```{r}

parcel_polygons_east <- st_read("K:\DataServices\Datasets\Town_Data\_Land_Parcel_Database\Processing\MassGIS_Parcel_Downloads\Statewide_parcels_SHP\L3_TAXPAR_POLY_ASSESS_EAST.shp")

parcel_polygons_west <- st_read("K:\DataServices\Datasets\Town_Data\_Land_Parcel_Database\Processing\MassGIS_Parcel_Downloads\Statewide_parcels_SHP\L3_TAXPAR_POLY_ASSESS_WEST.shp")

parcel_polygons <- rbind(parcel_polygons_east, parcel_polygons_west)
```

Land Parcel Database December 2023
```{r}
#Load full ccompressed assessors file of the land parcel database draft.
parcel_db <- read.csv(
  paste0(parcel_root, "LPDB_DRAFT_12.18.23.csv")
)  |> 
  dplyr::rename(
    Min_LUC = Min_LUC_Assign,
    Max_LUC = Max_LUC_Assign
  ) |> 
  left_join(
    keys,
    by = c("TOWN_ID" = "muni_id")
  )

#Read in parcel polygons.
parcel_polygons <- st_read(dsn = "C:/Users/bstanaway/Downloads/MassGIS_L3_Parcels_gdb.zip" , layer = "L3_TAXPAR_POLY")

# 1.0 Industrial Parcels
#Define land use code to filter on. 
ind_landuse_codes <- c("360","361","362","363","364","365","366","367","368","369",
  "400","401","402","403","404","410","411","412","413","420","421","423","424",
  "425","426","427","428","430","431","432","433","440","441","442","450","451",
  "452","230","231")

#Filter assessors records by industrial land use codes.
industrial_uses <- parcel_db |> 
  filter(
    Min_LUC %in% ind_landuse_codes
  )

#Join parcel polygons and filtered assessor's records.
#Include only the polygons with the exact LOC_ID match.
ind_shapes <- parcel_polygons |> 
  inner_join(
    industrial_uses,
    by = c("LOC_ID", "TOWN_ID")
  )

#Remove variables with names too long to be written to a shp file.
ind_shapes <- ind_shapes |> 
  select(
    names(ind_shapes[nchar(names(ind_shapes)) <= 9])
  )

#Convert the joined data.frame to a {sf} object to be written to a shp file.
ind_shapes <- st_as_sf(ind_shapes)

#Write shp file of the industrial parcels.
st_write(
  ind_shapes,
  "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/LPDB.Dec.23_Industrial.Parcels.shp"
)

rm(industrial_uses)
rm(ind_shapes)

gc()

#2.0 All MAPC Parcels
#Filter for parcel assessors records in MAPC
mapc_parcels <- parcel_db |> 
  filter(
    mapc == 1
  )

#Join parcel polygons and filtered assessor's records.
#Include only the polygons with the exact LOC_ID match.
mapc_shapes <- parcel_polygons |> 
  inner_join(
    mapc_parcels,
    by = c("LOC_ID", "TOWN_ID")
  )

#Remove variables with names too long to be written to a shp file.
mapc_shapes <- mapc_shapes |> 
  select(
    names(mapc_shapes[nchar(names(mapc_shapes)) <= 9])
  )

#Convert the joined data.frame to a {sf} object to be written to a shp file.
mapc_shapes <- st_as_sf(mapc_shapes)

#Write shp file of the industrial parcels.
st_write(
  mapc_shapes,
  "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/LPDB.Dec.23_MAPC.Parcels.shp"
)

rm(mapc_parcels)
rm(mapc_shapes)

gc()

#3.0 All MAPC Industrial Parcels
#Filter for parcel assessors records in MAPC
mapc.ind_parcels <- parcel_db |> 
  filter(
    mapc == 1
  ) |> 
  filter(
    Min_LUC %in% ind_landuse_codes
  )

#Join parcel polygons and filtered assessor's records.
#Include only the polygons with the exact LOC_ID match.
mapc.ind_shapes <- parcel_polygons |> 
  inner_join(
    mapc.ind_parcels,
    by = c("LOC_ID", "TOWN_ID")
  )

#Remove variables with names too long to be written to a shp file.
mapc.ind_shapes <- mapc.ind_shapes |> 
  select(
    names(mapc.ind_shapes[nchar(names(mapc.ind_shapes)) <= 9])
  )

#Convert the joined data.frame to a {sf} object to be written to a shp file.
mapc.ind_shapes <- st_as_sf(mapc.ind_shapes)

#Write shp file of the industrial parcels.
st_write(
  mapc.ind_shapes,
  "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/LPDB.Dec.23_MAPC.Ind.Parcels.shp"
)

rm(mapc.ind_parcels)
rm(mapc.ind_shapes)

gc()
```
