==> Industrial Landuse x Creative Economy/Scripts/2_Business.Data.Cleaning.Rmd <==
---
title: "2_Business.Data.Cleaning"
author: "Brandon Stanaway"
date: "2024-03-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

require(tidyverse)
require(tidycensus)
require(lubridate)
require(stringr)
require(janitor)
require(sf)
require(mapcdatakeys)
require(dplyr)
require(readxl)
require(tidygeocoder)
require(gt)
#source("C:/Project_Work/Arts_Culture/Industrial Landuse x Creative Economy/Scripts/municipal_name_fix.func.R")

#Load in data keys to add geographies to data later on.
keys <- mapcdatakeys::all_muni_data_keys |> 
  select(
    c(
      muni_name,
      muni_id,
      mapc,
      subrg_acr,
      cosub_5y21
    )
  ) |> 
  mutate(
    muni_name = stringr::str_to_title(muni_name)
  )

#Global options to remove scientific notation.
options(scipen = 999)
set.seed(351)

#Set root variables for local data ingestion
da_root <- "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Tabular/raw/"

```

1.0 Load DataAxle and A/C collected data
```{r}
#Load the 2023 DataAxls data
data.axle <- read.csv(
  paste0(da_root, "April_2023 - Massachusetts_New.csv")
)

#Load the 2023 RPA-input revised DataAxle data.
data.axle_revised <- read.csv(
  paste0(da_root, "Revised_DA_23/dataaxle_2023_revised.csv")
)

#Loads the Arts and Culture Spaces recorded in this airtable: https://airtable.com/appY1urvcJd9UmNyr
a.c_spaces <- read.csv(
  paste0(da_root, "AC Sites Airtable.csv")
)

#Loads Arts and Culture Spaces collected as part of the "Making Space for Art" project.
#A link to the airtable is here: https://airtable.com/appScRBzSMyNr011m/shrSi3WTESf2ARtK9/tblUIhUmEgbh8LY1h/viwMZVLjibMWEsHIl?blocks=show
makingspace_spaces <- read.csv(
  paste0(da_root, "MakingSpace AC Sites.csv")
)
```

2.1 Selecting Relevant Variables from DataAxle Data
```{r}
# 2023 edit - changing field names to reflect data axle fields
data.axle_clean <- data.axle |> 
  #Create an OBJECT_ID variable
  mutate(
    OBJECT_ID = seq.int(nrow(data.axle))
  ) |> 
  #Select the variables needed for this analysis.
  dplyr::select(OBJECT_ID, CONAME, STADDR, STCITY, STATE, ZIP, STCODE, 
                CNTYCD, CENSUS, BLKGRP, LATITUDE, LONGITUDE, NAICS, NAICSD, 
                EMPSIZ, LOCEMP, SALVOL, SLSVDT, ZPOPCD, WEALTH, BBUSIN, 
                MBUSIN, SBUSIN, HOME, INDFRM, FEMOWN
  ) |> 
  #Creat variables that represent the 4- and 6-digit versions of NAICS codes
  mutate(
    NAICS.4 = as.character(substr(as.character(NAICS), 1, 4)),
    NAICS.6 = as.character(substr(as.character(NAICS), 1, 6))
  ) |> 
  #Filter for observations with non-N/A employment + are in MA.
  filter(
    !is.na(LOCEMP) & STCODE == 25
  ) 
```


2.2 Harmonizing Town Names
```{r}
#Fix incorrect municipality names in the data
data.axle_clean <- data.axle_clean |> 
    mutate(
      STCITY_CLEAN = case_when(
        #Manchester
        STCITY %in% c("Manchester-by-the-Sea","Manchestr By Sea") ~ "Manchester",
        #Foxborough
        STCITY %in% c("Foxboro") ~ "Foxborough",
        #North Attleborough
        STCITY %in% c("North Attleboro") ~ "North Attleborough",
        #Boston
        STCITY %in% c("Roxbury", "Roxbury Crossing", "Mission Hill",
                             "Dorchester", "Grove Hall", "Dorchester Ctr",
                             "Uphams's Corner", "Mattapan", "South Boston",
                             "East Boston", "Charlestown", "Jamaica Plain",
                             "Roslindale", "West Roxbury", "Allston",
                             "Brighton", "Hyde Park") ~ "Boston",
        #Arlington
        STCITY %in% c("East Arlington", "Arlington Hts") ~ "Arlington",
        #Cambridge
        STCITY %in% c("North Cambridge", "East Cambridge") ~ "Cambridge",
        #Newton
        STCITY %in% c("Newtonville", "Newton Center", "Newton Highlands",
                             "Newton Lower Fls", "Newton Lower Falls", 
                             "Newton Upper Fls", "Newton Upper Falls",
                             "West Newton", "Auburndale", "Chestnut Hill",
                             "Waban", "Nonatum") ~ "Newton",
        #Easton
        STCITY %in% c("North Easton", "South Easton") ~ "Easton",
        #Weymouth
        STCITY %in% c("East Weymouth", "South Weymouth", "North Weymouth") ~ "Weymouth",
        #Falmouth
        STCITY %in% c("Teaticket", "North Falmouth", "East Falmouth",
                             "West Falmouth", "Woods Hole", "Waquoit") ~ "Falmouth",
        #Quincy
        STCITY %in% c("North Quincy", "Wollaston") ~ "Quincy",
        #Barnstable
        STCITY %in% c("Hyannis", "Centerville", "Cotuit", "Cummaquid", 
                             "Hyannis Port", "Marstons Mills", "Osterville", "West Hyannisport",
                             "West Barnstable") ~ "Barnstable",
        #Somerville
        STCITY %in% c("West Somerville") ~ "Somerville",
        #Plymouth
        STCITY %in% c("White Horse Bch", "Manomet") ~ "Plymouth",
        #Sandwich
        STCITY %in% c("East Sandwich", "Forestdale") ~ "Sandwich",
        #Wellfleet
        STCITY %in% c("South Wellfleet") ~ "Wellfleet",
        #Amherst
        STCITY %in% c("North Amherst") ~ "Amherst",
        #Uxbridge
        STCITY %in% c("Noth Uxbridge") ~ "Uxbridge",
        #Great Barrington
        STCITY %in% c("Housatonic") ~ "Great Barrington",
        #Tisbury
        STCITY %in% c("Vineyard Haven") ~ "Tisbury",
        #Scituate
        STCITY %in% c("Greenbush", "Humarock", "North Scituate") ~ "Scituate",
        #Newbury
        STCITY %in% c("Byfield") ~ "Newbury",
        #Walpole
        STCITY %in% c("East Walpole", "South Walpole") ~ "Walpole",
        #Grafton
        STCITY %in% c("North Grafton", "South Grafton") ~ "Grafton",
        #Colrain
        STCITY %in% c("Shattuckville") ~ "Colrain",
        #Russell
        STCITY %in% c("Woronoco") ~ "Russell",
        #Milton
        STCITY %in% c("Milton Village") ~ "Milton",
        #Sheffield
        STCITY %in% c("Ashley Falls") ~ "Sheffield",
        #North Attleborough
        STCITY %in% c("Attleboro Falls") ~ "North Attleborough",
        #Sutton
        STCITY %in% c("Manchaug", "Wilkinsonville") ~ "Sutton",
        #Hatfield
        STCITY %in% c("West Hatfield", "North Hatfield") ~ "Hatfield",
        #Douglas
        STCITY %in% c("East Douglas") ~ "Douglas",
        #Orleans
        STCITY %in% c("East Orleans") ~ "Orleans",
        #Attleboro
        STCITY %in% c("South Attleboro") ~ "Attleboro",
        #Templeton
        STCITY %in% c("Baldwinville", "East Templeton") ~ "Templeton",
        #Charlton
        STCITY %in% c("Charlton Depot", "Charlton City") ~ "Charlton",
        #Freetown
        STCITY %in% c("Assonet", "East Freetown") ~ "Freetown",
        #Millbury
        STCITY %in% c("West Millbury") ~ "Millbury",
        #Harwich
        STCITY %in% c("West Harwich", "Harwich Port", "East Harwich") ~ "Harwich",
        #Dennis
        STCITY %in% c("South Dennis", "West Dennis", "East Dennis",
                             "Dennis Port") ~ "Dennis",
        #Montague
        STCITY %in% c("Lake Pleasant", "Millers Falls", "Turners Falls") ~ "Montague",
        #Yarmouth
        STCITY %in% c("Bass River", "West Yarmouth", "Yarmouth Port",
                             "South Yarmouth") ~ "Yarmouth",
        #Marshfield
        STCITY %in% c("Marshfield Hills", "North Marshfield", "Ocean Bluff",
                             "Brant Rock", "Green Harbor") ~ "Marshfield",
        #Norwell
        STCITY %in% c("Accord") ~ "Norwell",
        #Brookline
        STCITY %in% c("Brookline Vlg") ~ "Brookline",
        #Gosnold
        STCITY %in% c("Cuttyhunk") ~ "Gosnold",
        #Williamsburg
        STCITY %in% c("Haydenville") ~ "Williamsburg",
        #Medford
        STCITY %in% c("West Medford") ~ "Medford",
        #Whately
        STCITY %in% c("West Whately") ~ "Whately",
        #Otis
        STCITY %in% c("East Otis") ~ "Otis",
        #Wellesley
        STCITY %in% c("Wellesley Hills", "Babson Park") ~ "Wellesley",
        #Agawam
        STCITY %in% c("Feeding Hills") ~ "Agawam",
        #Hardwick
        STCITY %in% c("Gilbertville", "South Harwich") ~ "Hardwick",
        #Bourne
        STCITY %in% c("Buzzards Bay", "Cataumet", "Monument Beach",
                             "Pocasset", "Sagamore", "Sagamore Beach") ~ "Bourne",
        #Westport
        STCITY %in% c("Westport Point") ~ "Westport",
        #Eastham
        STCITY %in% c("North Eastham") ~ "Eastham",
        #Wareham
        STCITY %in% c("East Wareham", "West Wareham", "Onset") ~ "Wareham",
        #Lanesborough
        STCITY %in% c("Berkshire") ~ "Lanesborough",
        #Palmer
        STCITY %in% c("Bondsville", "Thorndike", "Three Rivers") ~ "Palmer",
        #Haverhill
        STCITY %in% c("Bradford") ~ "Haverhill",
        #Pembroke
        STCITY %in% c("Bryantville") ~ "Pembroke",
        #Norton
        STCITY %in% c("Chartley") ~ "Norton",
        #Leicester
        STCITY %in% c("Cherry Valley", "Rochdale") ~ "Leicester",
        #Florida
        STCITY %in% c("Drury") ~ "Florida",
        #Taunton
        STCITY %in% c("East Taunton") ~ "Taunton",
        #East Bridgewater
        STCITY %in% c("Elmwood") ~ "East Bridgewater",
        #Sturbridge
        STCITY %in% c("Fiskdale") ~ "Sturbridge",
        #Northampton
        STCITY %in% c("Florence", "Leeds") ~ "Northampton",
        #Danvers
        STCITY %in% c("Hathorne") ~ "Danvers",
        #Springfield
        STCITY %in% c("Indian Orchard") ~ "Springfield",
        #Holden
        STCITY %in% c("Jefferson") ~ "Holden",
        #Lenox
        STCITY %in% c("Lenox Dale") ~ "Lenox",
        #Stockbrodge
        STCITY %in% c("Glendale") ~ "Stockbridge",
        #Chilmark
        STCITY %in% c("Menemsha") ~ "Chilmark",
        #New Marlborough
        STCITY %in% c("Mill River") ~ "New Marlborough",
        #Hanson
        STCITY %in% c("Monponsett") ~ "Hanson",
        #Gill
        STCITY %in% c("Mt Hermon") ~ "Gill",
        #Needham
        STCITY %in% c("Needham Heights") ~ "Needham",
        #Billerica
        STCITY %in% c("North Billerica") ~ "Billerica",
        #Carver
        STCITY %in% c("North Carver") ~ "Carver",
        #Chatham
        STCITY %in% c("North Chatham") ~ "Chatham"
      )
    ) |> 
  mutate(
    STCITY_CLEAN = coalesce(STCITY_CLEAN, STCITY)
  ) |> 
  #Join MAPC data keys to the data axle data to filter by relevant geographies
  left_join(
    keys,
    c("STCITY_CLEAN" = "muni_name")
  ) |> 
  select(
    -c(
      STCITY
    )
  )
```

2.3 Geocoding Addresses
```{r}
#Create a concatenated address field: Street Address, City, State, ZIP Code
data.axle_clean.geocoded <- data.frame()
muni_id <- unique(data.axle_clean$muni_id)

for (i in muni_id){
  
  tmp <- data.axle_clean |> 
    filter(
      muni_id == i
    ) |> 
  mutate(
    #Convert ZIP to string + add leading zero to make ZIP codes consistently five
    #digits.
    ZIP = as.character(str_pad(ZIP, width = 5, side = c("left"), pad = 0)),
    #Create a complete address column.
    COMPLETE_ADDR = str_c(STADDR, STCITY_CLEAN, STATE, ZIP, sep = " ")
  ) |> 
  #Call the geocoder
  tidygeocoder::geocode(
    address = COMPLETE_ADDR,
    method = "arcgis"
  )
  
  data.axle_clean.geocoded <- bind_rows(data.axle_clean.geocoded, tmp)
  
  print(paste0("Finished Geocoding Municipality:", i))
  
  rm(tmp)
  
}
```

2.4 Filtering Arts Businesses
```{r}
# From Appendix D (NEFA Core Industries by Creative Category) of 2017 NEFA Report "The Jobs in New England's Creative Economy and Why They Matter"
# Creation of industry groups by six digit NAICS code (assign to establishments with ifelse function)
arch.and.design <- c('541310', '541320', '541340', '541410', '541420', '541430', '541490')
#art.electronic.retail <- c('443142', '448310', '451130', '451140', '451211', '453920', '812921', '812922')
art.retail <- c('448310', '451130', '451140', '451211', '453920', '812921', '812922') # no electronics-related retail
art.arch.manu <- c('332323', '337212', '339910', '339992')
culture.pres <- c('712110', '712120', '712130', '712190')
mach.comm.manu <- c('333244', '334310', '334614')
marketing <- c('541810', '541830', '541840', '541850')
mat.manu <- c('325992', '327110', '327212', '339940')
media <- c('515111', '515112', '515120', '515210', '517110', '519110', '519120', '519130', '532230')
motion.pic.teleprod <- c('512110', '512120', '512131', '512132', '512191', '512199')
music.rec <- c('512210', '512220', '512230', '512240', '512290')
printing <- c('323111', '323113', '323117', '323120')
publishing <- c('511110', '511120', '511130', '511191', '511199')
visual.music.perf.arts <- c('541921', '541922', '611610', '711110', '711120', '711130', '711190', '711510')
wholesale.art.store <- c('423410', '423940', '424110', '424920')

#Create a full list of arts related businesses.
arts_sectors <- c(arch.and.design, art.retail, art.arch.manu, culture.pres,mach.comm.manu,marketing,
                  mat.manu, media, motion.pic.teleprod, music.rec, printing, publishing, visual.music.perf.arts,
                  wholesale.art.store)

#Filter for the arts related businesses.
data.axle_arts.spaces <- data.axle_clean |> 
  filter(
    NAICS.6 %in% arts_sectors
  )
```

2.5 Filter for MAPC geography
```{r}
#Join the MAPC data keys to the arts spaces data.
data.axle_arts.spaces_mapc <- data.axle_arts.spaces |> 
  filter(
    mapc == 1
  )

```

2.5.1 Geocode Arts Spaces
```{r}
#Create a concatenated address field: Street Address, City, State, ZIP Code
data.axle_arts.spaces_mapc.geocoded <- data.frame()
muni_id <- unique(data.axle_arts.spaces_mapc$muni_id)

for (i in muni_id){
  
  tmp <- data.axle_arts.spaces_mapc |> 
    filter(
      muni_id == i
    ) |> 
    mutate(
    #Convert ZIP to string + add leading zero to make ZIP codes consistently five
    #digits.
    ZIP = as.character(str_pad(ZIP, width = 5, side = c("left"), pad = 0)),
    #Create a complete address column.
    COMPLETE_ADDR = str_c(STADDR, STCITY_CLEAN, STATE, ZIP, sep = " ")
    ) |> 
    #Call the geocoder
    tidygeocoder::geocode(
      address = COMPLETE_ADDR,
      method = "arcgis"
    )
  
  data.axle_arts.spaces_mapc.geocoded <- bind_rows(data.axle_arts.spaces_mapc.geocoded, tmp)
  
  print(paste0("Finished Geocoding Municipality:", i))
  
  rm(tmp)
}
```


2.5.2 Write cleaned Business data
```{r}
write.csv(
  data.axle_arts.spaces_mapc.geocoded,
  "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Tabular/intermediary/DA.23_ArtsSpaces.MAPC.cleaned.geocoded.csv",
  row.names = F
)

saveRDS(
  data.axle_arts.spaces_mapc.geocoded,
  file = "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Tabular/intermediary/DA.23_ArtsSpaces.MAPC.cleaned.geocoded.rds"
)
```

2.6 MAPC Summary Statistics
```{r}
output_path <- "C:/Project_Work/Arts_Culture/Industrial Landuse x Creative Economy/Data/Outputs/"

#Number of Arts Businesses by Municipality
arts.biz_by.town <- data.axle_arts.spaces_mapc |> 
  group_by(
    STCITY_CLEAN
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(businesses)
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.town,
  paste0(output_path, "arts.business_by_muni.csv")
)

#Create a table demonstrating the top 10 municipalities by number of arts businesses
tbl.1 <- arts.biz_by.town |>
  dplyr::slice(1:10) |> 
  gt() |> 
  tab_header(
    title = "NEFA Arts Businesses by Municipality - MAPC"
  ) |> 
  cols_label(
    STCITY_CLEAN = "Municipality",
    businesses = "Businesses",
    employment = "Employment",
    emp.by.biz = "Avg. Employment Per Business"
  ) |> 
  fmt_number(
    columns = c("businesses", "employment"),
    sep_mark = ",",
    decimals = 0
  )

tbl.1 |> 
  gtsave(
    filename = paste0(output_path, "arts.biz_by.town.html")
  )

#Analysis
#99 of MAPC's 101 municipalities are home to at least one arts business.

#=======================================================================================
#Number of Arts Businesses by 6-Digit NAICS Code
arts.biz_by.naics <- data.axle_arts.spaces_mapc |> 
  group_by(
    NAICSD,
    NAICS.6
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(businesses)
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.naics,
  paste0(output_path, "arts.business_by_NAICS.csv")
)

#Create a table demonstrating the top 10 NAICS Codes by number of arts businesses
tbl.2 <- arts.biz_by.naics |>
  dplyr::slice(1:10) |> 
  gt() |> 
  tab_header(
    title = "NEFA Art Businesses by NAICS Code - MAPC"
  ) |> 
  cols_label(
    NAICSD = "NAICS",
    NAICS.6 = "6-Digit NAICS Code",
    businesses = "Businesses",
    employment = "Employment",
    emp.by.biz = "Avg. Employment Per Business"
  ) |> 
  fmt_number(
    columns = c("businesses", "employment"),
    sep_mark = ",",
    decimals = 0
  )

tbl.2 |> 
  gtsave(
    filename = paste0(output_path, "arts.biz_by.naics.html")
  )

#Analysis
#49 different types of arts business are represented in the MAPC region.


#==============================================================================
#Number of Arts businesses by municipality and 6-Digit NAICS Code
arts.biz_by.town.by.naics <- data.axle_arts.spaces_mapc |> 
  group_by(
    STCITY_CLEAN,
    NAICSD,
    NAICS.6
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(desc(STCITY_CLEAN))
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.town.by.naics,
  paste0(output_path, "arts.business_by_muni_by_NAICS.csv")
)

#Analysis
#

```

2.6 Summary Statistics Maps
```{r}
```

3.0 Compare DataAxle businesses to MAPC-collected Arts data
3.1 Define the Haversine formula (measures the distance between two points on a sphere)
```{r haversine-formula}
# Harversine formula implementation

haversine = function(long1, lat1, long2, lat2, round=3)
{
  # convert to radians
  long1 = long1 * pi / 180
  lat1  = lat1  * pi / 180
  long2 = long2 * pi / 180
  lat2  = lat2  * pi / 180
  
  R = 6371 # Earth mean radius in km

  a = sin((lat2 - lat1)/2)^2 + cos(lat1) * cos(lat2) * sin((long2 - long1)/2)^2
  d = R * 2 * asin(sqrt(a))

  return( round(d,round) ) # distance in km
}
```

3.2 Create a distance matrix between the DataAxle data and Making Space data
```{r nested-for-loop, eval=FALSE}
# Nested for loop distance calculation implementation

dist = matrix(NA, nrow(data.axle_arts.spaces_mapc), nrow(makingspace_spaces))
for(i in 1:nrow(data.axle_arts.spaces_mapc)){
  for(j in 1:nrow(makingspace_spaces)){
    dist[i,j] = haversine(data.axle_arts.spaces_mapc$LONGITUDE[i], data.axle_arts.spaces_mapc$LATITUDE[i],
                          makingspace_spaces$long[j], makingspace_spaces$lat[j])
  }
}
```
