==> Industrial Landuse x Creative Economy/Scripts/3_Parcels.Businesses.Spatial.Join.Rmd <==
---
title: "3_Parcels.Businesses.Spatial.Join"
author: "Brandon Stanaway"
date: "2024-03-28"
output: html_document
---

# 0.0 Set up
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(tidycensus)
library(lubridate)
library(stringr)
library(janitor)
library(sf)
library(mapcdatakeys)
library(dplyr)
library(readxl)
library(ggplot2)
library(spdep)
library(ggrepel)
library(gt)
#source("C:/Project_Work/Arts_Culture/Industrial Landuse x Creative Economy/Scripts/municipal_name_fix.func.R")

#Load in data keys to add geographies to data later on.
keys <- mapcdatakeys::all_muni_data_keys |> 
  select(
    c(
      muni_name,
      muni_id,
      mapc,
      subrg_acr
    )
  ) |> 
  mutate(
    muni_name = stringr::str_to_title(muni_name)
  )

#Define land use code to filter on. 
ind_landuse_codes <- c("360","361","362","363","364","365","366","367","368","369",
  "400","401","402","403","404","410","411","412","413","420","421","423","424",
  "425","426","427","428","430","431","432","433","440","441","442","450","451",
  "452","230","231")

#Global options to remove scientific notation.
options(scipen = 999)
set.seed(351)

#Set root variables for local data ingestion
#Parcels shapefile
parcels_root <- "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/"

#Business data 
biz_root <- "C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Tabular/intermediary/"

#Set output path for data outputs
output_path <- "C:/Project_Work/Arts_Culture/Industrial Landuse x Creative Economy/Data/Outputs/"

#Set common projecrions + crs for conversion.
mass_mainland<-"+proj=lcc +lat_1=42.68333333333333 +lat_2=41.71666666666667 +lat_0=41 +lon_0=-71.5 +x_0=200000 +y_0=750000 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
lat_lon_CRS <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"

#Load Land Use Code Crosswalk
LUC_Descriptions <- read.csv(
  "C:/Project_Work/Local_Data/Parcel_DB/comprehensive_adj_use.code_xwalk.csv"
)

#reading in MAPC's municipalities shapefile
mapc_municipalities <- st_read("https://services.arcgis.com/c5WwApDsDjRhIVkH/arcgis/rest/services/MA_Municipalities/FeatureServer/0/query?outFields=*&where=1%3D1&f=geojson") |>   
  select(town_id, geometry) |> 
  rename(muni_id = town_id) |> 
  left_join(
    keys,
    by = c("muni_id")
  )

#Reading in some municipal boundaries
nbhds <- st_read("C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/partnercity_nbhds.shp")

#Create centoids of the neighborhood shapefiles to zoom into the nbhd in maps.
nbhds_centroids <- st_point_on_surface(nbhds)

#Reading in the Fabrication (Fab) Zoning District for Somerville analysis
fab_dist <- st_read("C:/Project_Work/Local_Data/Arts_Culture/Industrial_Land_Use/Data/Spatial/Zoning.shp") |> filter(ZoneCode == "FAB")

```

# 1.0 Load relevant data
```{r}
#1.1 Parcels
#Load the parcel polygons
mapc_parcels <- st_read(
  paste0(parcels_root, "LPDB.Dec.23_MAPC.Parcels.shp")
)

#1.2 Businesses
#Load the business data
#Convert business point data to a a spatial object (sf object)
mapc_biz.spatial <- readRDS(
  paste0(biz_root, "DA.23_ArtsSpaces.MAPC.cleaned.geocoded.rds")
) |> 
  rowwise() |> 
  mutate(
    lat = tidyr::replace_na(lat, LATITUDE),
    long = tidyr::replace_na(long, LONGITUDE)
  ) |> 
  #Converting the non-spatial data into a spatial object.
  st_as_sf(
    #coords = c("LONGITUDE", "LATITUDE"),
    coords = c("long","lat"),
    crs = st_crs(lat_lon_CRS),
    remove = FALSE
  ) |>
  #Setting CRS to match the polygon's CRS.
  st_transform(
    mass_mainland,
    crs = st_crs(mapc_parcels)
  )
```

# 2.0 Land Use Code Descriptions + Symbology
```{r}
mapc_parcels <- mapc_parcels |> 
    mutate(
    USE_CODE_SYMB = case_when(
      Min_LUC %in% c("130","131","132") ~ "Residential Vacant",
      Min_LUC %in% c("101") ~ "Single Family Residence",
      Min_LUC %in% c("104","105") ~ "Two or Three Family Residences",
      Min_LUC %in% c("102") ~ "Condominium",
      Min_LUC %in% c("111", "112") ~ "Apartments with Four or more units",
      Min_LUC %in% c("121","122","123","124","125") ~ "Group Quarters",
      Min_LUC %in% c("109","140","107","108","100","110","113","114","115","116","117","118","119","120",
                        "126","127","133","136","137","138","13C","150","160","167","170","171","172","173",
                        "176","180","199","106","103") ~ "Residential Other",
      Min_LUC %in% c("0xx") ~ "Mixed Use",
      Min_LUC %in% c("390","391","392","393") ~ "Commercial Vacant",
      Min_LUC %in% c("300","301","302","303","304","305","306","310","311","312","313","314","315","316",
                        "317","318","321","322","323","324","325","326","330","331","332","333","334","335",
                        "336","337","338","340","342","350","351","352","353","354","355","356","360","361",
                        "362","363","364","365","366","367","368","369","370","371","372","373","374","375",
                        "376","377","380","381","382","383","384","385","386","387","388","389","307","308",
                        "309","319","31l","320","327","328","329","339","33l","343","344","345","346","347",
                        "348","349","34R","378","379","394","395","396","399")~ "Commercial",
      Min_LUC %in% c("440","441","442") ~ "Industrial Vacant",
      Min_LUC %in% c("400","401","402","403","404","410","411","412","413","420","421","422","423","424",
                        "425","426","427","428","430","431","432","433","450","451","452","405","406","407",
                        "408","409","414","415","417","434","435","436","443","444","445","446","465","490",
                        "499") ~ "Industrial",
      Min_LUC %in% c("201","202","210","211","220","221","231","261","271","272","273","274","276","277",
                        "278","279","281","285","290") ~ "Open Space",
      Min_LUC %in% c("601","602","710","711","712","713","714","715","716","717","718","719","720","801",
                        "802","803","804","805","806","807","808","809","810","811","812","813","814","815") ~ "Chapter 61 or 61A or 61B Property",
      Min_LUC %in% c("237","292","600","610","670","700","701","722","730","733","734","736","737","760",
                        "800") ~ "Other Agricultural or Recreational",
      Min_LUC %in% c("930","932","933","936","938","973","974","975","980","982","988","991") ~ "Municipal Vacant",
      Min_LUC %in% c("900","901","910","911","912","913","914","915","916","917","918","919","920","921",
                        "922","923","924","925","926","927","928","929","931","934","935","937","939","981",
                        "985","970","971","972","989","992","993","994") ~ "Federal State or Municipal",
      Min_LUC %in% c("946","950") ~ "Institutional or Exempt Vacant",
      Min_LUC %in% c("940","941","942","943","944","945","945","947","951","952","953","954","955","956",
                        "957","957","958","959","960","961","962","995","996","997","990") ~ "Institutional or Exempt",
      Min_LUC %in% c("902","903","904","905","906","907","908","909","948","949","963","965","966","968",
                        "969","976","977","978","979","983","986","987","998","999") ~ "Tax Exempt Other"
    ),
    USE_CODE_SYMB = replace_na(USE_CODE_SYMB, "Unknown")
  )
```

# 2.1 Putting the points into the polygons
```{r}
#Testing that the points are projected similar to the polygons.
#Using Stow as an example because it's small + has only a few points
biz_test <- mapc_biz.spatial |> 
  filter(
    STCITY_CLEAN == "Stow"
  )

parcel_test <- mapc_parcels |> 
  filter(
    CITY == "Stow"
  )

ggplot()+
  geom_sf(data = parcel_test)+
  geom_sf(data = biz_test) +
  theme_minimal()

#Hell yeah - the points are in the polygons!
#Lets remove the test dataframes
rm(biz_test, parcel_test)

#Now lets do the point in polygon join for the full dataset
biz_in_parcel <- st_join(mapc_biz.spatial, mapc_parcels, join = st_within)
biz_in_muni <- st_join(mapc_biz.spatial, mapc_municipalities, join = st_within)
```

# 2.2 QA/QC SMall Muni Check
```{r}
address_check <- biz_in_parcel |> 
  select(
    Min_LUC,
    CONAME,
    STADDR,
    STCITY_CLEAN,
    LOC_ID,
    ADDR_NUM,
    FULL_STR,
    CITY
  ) |> 
  filter(
    STCITY_CLEAN == "Stow"
  ) |>
  mutate(
    STADDR_NUM = word(STADDR, start = 1, sep = " "),
    addr_num_check = if_else(ADDR_NUM == STADDR_NUM, 1 , 0)
  )

joyleduc_parcel <- mapc_parcels |> 
  filter(
    LOC_ID == "M_201269_909233"
  )

joyleduc_biz <- mapc_biz.spatial |> 
  filter(
    CONAME == "Joy Leduc Photography LLC"
  )

ggplot()+
  geom_sf(data = joyleduc_parcel)+
  geom_sf(data = joyleduc_biz) +
  theme_minimal()

#Multi_address, single loc_id parcels dont have all address information. Make an address match an issue.
#Data Axle does not have full addresses some of the time.
  
```

# 2.3 Medium Muni Check
```{r}
#Dataset to check address matching
address_check <- biz_in_parcel |> 
  select(
    Min_LUC,
    CONAME,
    STADDR,
    STCITY_CLEAN,
    LOC_ID,
    ADDR_NUM,
    FULL_STR,
    CITY
  ) |> 
  filter(
    STCITY_CLEAN == "Waltham"
  ) |>
  mutate(
    STADDR_NUM = word(STADDR, start = 1, sep = " "),
    addr_num_check = if_else(ADDR_NUM == STADDR_NUM, 1 , 0)
  )

#First Biz
hp_biz <- mapc_biz.spatial |> 
  filter(
    CONAME == "Harvard Printing" & STCITY_CLEAN == "Waltham"
  )

waltham_parcels <- mapc_parcels |> 
  filter(
    CITY == "Waltham" & (FULL_STR == "ELM ST" | FULL_STR == "MOODY ST")
  )

ggplot() +
  geom_sf(data = waltham_parcels, aes(fill = FULL_STR))+
  geom_sf(data = hp_biz) +
  theme_minimal()

#Data axle list this business to be located at 36 Elm Street. The point falls at 362 Moody Street. 
#Data Axle has LAT/LON that do not match the address given in the data.
#FIXED via geocoding!

#Second Biz
rsp_biz <- mapc_biz.spatial |> 
  filter(
    CONAME == "Hajian Architects Inc" & STCITY_CLEAN == "Waltham"
  )

waltham_parcels <- mapc_parcels |> 
  filter(
    CITY == "Waltham" & (FULL_STR == "DEXTER ST" | FULL_STR == "DEXTER AVE")
  )

ggplot() +
  geom_sf(data = waltham_parcels, aes(fill = FULL_STR))+
  geom_sf(data = rsp_biz) +
  theme_minimal()

#Here we have an example of the parcel data not having am address (or parcel) that matches the address of the business (56 Dexter Ave vs 32 Dexter Street). That said, 56 Dexter Ave was geocoded to the appropriate location (no 56 Dexter Ave exists.)
```
# 2.4 Pembroke Missing Arts Businesses Visualization
```{r}
#Select the relevant information from the point in polygon join
address_check <- biz_in_parcel |> 
  select(
    Min_LUC,
    CONAME,
    STADDR,
    STCITY_CLEAN,
    LOC_ID,
    ADDR_NUM,
    FULL_STR,
    CITY
  ) |> 
  filter(
    STCITY_CLEAN == "Pembroke"
  ) |>
  mutate(
    STADDR_NUM = word(STADDR, start = 1, sep = " "),
    addr_num_check = if_else(ADDR_NUM == STADDR_NUM, 1 , 0)
  )

#Select the businesses on the street we are interested in.
#In this case it's Corporate Park Drive; where Soundcheck Studios is located.
hp_biz <- mapc_biz.spatial |> 
  filter(
    stringr::str_detect(STADDR, "Corporate Park Dr") & STCITY_CLEAN == "Pembroke"
  ) |> 
  #Removing a duplicate business
  filter(
    OBJECT_ID != "199458"
  )

#Filter for parcel polygons on Corporate Park Drive
waltham_parcels <- mapc_parcels |> 
  filter(
    CITY == "Pembroke" & (FULL_STR == "CORPORATE PARK DRIVE")
  )

#Filter for the parcel on which Soundcheck Studios is located to symbolize it in the map.
missing_biz_parcel <- mapc_parcels |> 
  filter(
    CITY == "Pembroke" & FULL_STR == "CORPORATE PARK DRIVE" & ADDR_NUM == "150"
  )

#Map of Corporate Park Drive (Pembroke) parcels and arts businsesses
pembroke.map <- ggplot() +
  geom_sf(data = waltham_parcels, fill = NA)+
  geom_sf(data = hp_biz, color = "blue") +
  geom_label_repel(
    data = hp_biz,
    aes(label = CONAME, geometry = geometry),
    stat = "sf_coordinates",
    min.segment.length = 0
  ) +
  geom_sf(data = missing_biz_parcel, fill = NA, color = "red", linewidth = 1.0) +
  labs(
    title = "DataAxle is Missing Established Arts Businesses \n(Soundcheck Studios) In Pembroke",
    subtitle = "Corporate Park Drive Parcels and Arts Businesses (blue) \n150 Corporate Park Drive (red)",
    y = "",
    x = ""
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

pembroke.map

ggsave(
  filename = "Pembroke_map.jpeg",
  plot = pembroke.map,
  device = "jpeg",
  path = output_path
)
```
# 2.4 Norwood Missing Arts Businesses Visualization
```{r}
#Select the relevant information from the point in polygon join
address_check <- biz_in_parcel |> 
  select(
    Min_LUC,
    CONAME,
    STADDR,
    STCITY_CLEAN,
    LOC_ID,
    ADDR_NUM,
    FULL_STR,
    CITY
  ) |> 
  filter(
    STCITY_CLEAN == "Norwood"
  ) |>
  mutate(
    STADDR_NUM = word(STADDR, start = 1, sep = " "),
    addr_num_check = if_else(ADDR_NUM == STADDR_NUM, 1 , 0)
  )

#Select the businesses on the street we are interested in.
#In this case it's Morse Street; where Norwood Space Center is located.
hp_biz <- mapc_biz.spatial |> 
  filter(
    stringr::str_detect(STADDR, "Morse St") & STCITY_CLEAN == "Norwood"
  )

#Filter for parcel polygons on Morse Street
waltham_parcels <- mapc_parcels |> 
  filter(
    CITY == "Norwood" & FULL_STR == "MORSE ST"
  )

#Filter for the parcel on which Norwood Space Center is located to symbolize it in the map.
missing_biz_parcel <- mapc_parcels |> 
  filter(
    CITY == "Norwood" & FULL_STR == "MORSE ST" & ADDR_NUM == 83
  )

#Map of Morse Street (Norwood) parcels and arts businesses
norwood.map <- ggplot() +
  geom_sf(data = waltham_parcels, fill = NA)+
  geom_sf(data = hp_biz, color = "blue") +
  geom_label_repel(
    data = hp_biz,
    aes(label = CONAME, geometry = geometry),
    stat = "sf_coordinates",
    min.segment.length = 0
  ) +
  geom_sf(data = missing_biz_parcel, fill = NA, color = "red", linewidth = 1.0) +
  labs(
      title = "DataAxle is Missing Established Arts Spaces \n(Norwood Space Center) In Norwood",
    subtitle = "Morse Street Parcels and Arts Businesses (blue) \n83 Morse Street Parcel is Missing \nShould be where the leftmost points are located",
    y = "",
    x = ""
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

norwood.map

ggsave(
  filename = "Norwood_map.jpeg",
  plot = norwood.map,
  device = "jpeg",
  path = output_path
)
```

# 2.5 Creating a Parcel-level Count of Businesses and Employment
```{r}
#Table to rejoin to parcel geometries for spatial visualization
Parcel_Biz <- biz_in_parcel |> 
  group_by(
    LOC_ID
  ) |> 
  summarise(
    biz = n(),
    emp = sum(LOCEMP),
    epb = emp/biz
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry()

# Table of 
Industrial.Parcel_Biz <- biz_in_parcel |> 
  mutate(USE_CODE_SYMB = case_when(
    USE_CODE_SYMB == "Industrial Vacant" ~ "Industrial",
    USE_CODE_SYMB == "Commercial Vacant" ~ "Commercial",
    USE_CODE_SYMB == "Single Family Residence" ~ "Residential",
    USE_CODE_SYMB == "Two or Three Family Residences" ~ "Residential",
    USE_CODE_SYMB == "Residential Other" ~ "Residential",
    USE_CODE_SYMB == "Residential Vacant" ~ "Residential",
    USE_CODE_SYMB == "Apartments with Four or more units" ~ "Residential",
    .default = USE_CODE_SYMB
  )) |>
  group_by(
    muni_name,
    USE_CODE_SYMB
  ) |> 
  summarise(
    arts.biz = n(),
    arts.emp = sum(LOCEMP),
    arts.epb = arts.emp/arts.biz
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry()

# 
tabular_parcels <- mapc_parcels |> 
  st_drop_geometry() |> 
  mutate(USE_CODE_SYMB = case_when(
    USE_CODE_SYMB == "Industrial Vacant" ~ "Industrial",
    USE_CODE_SYMB == "Commercial Vacant" ~ "Commercial",
    USE_CODE_SYMB == "Single Family Residence" ~ "Residential",
    USE_CODE_SYMB == "Two or Three Family Residences" ~ "Residential",
    USE_CODE_SYMB == "Residential Other" ~ "Residential",
    USE_CODE_SYMB == "Residential Vacant" ~ "Residential",
    USE_CODE_SYMB == "Apartments with Four or more units" ~ "Residential",
    .default = USE_CODE_SYMB
  )) |> 
  group_by(muni_name, USE_CODE_SYMB) |> 
  summarise(
    num_parcels = n()
  ) |> 
  ungroup() |> 
  group_by(
    muni_name
  ) |> 
  mutate(
    tot_parcels = sum(num_parcels)
  ) |> 
  ungroup() |> 
  mutate(
    pct_parcels = num_parcels/tot_parcels
  )

x <- left_join(
  tabular_parcels,
  Industrial.Parcel_Biz,
  by = c("muni_name", "USE_CODE_SYMB")
) |> 
  mutate(
    across(everything(), ~replace_na(.x, 0)),
    pct_Parcels_w_Arts.Biz = arts.biz/num_parcels
  )

# Industrial
x.industrial <- x |> 
  filter(
    USE_CODE_SYMB %in% c("Industrial")
  ) |> 
  mutate(
    rank = (pct_Parcels_w_Arts.Biz - min(pct_Parcels_w_Arts.Biz))/ (max(pct_Parcels_w_Arts.Biz) - min(pct_Parcels_w_Arts.Biz))
  ) |> 
  select(
    muni_name,
    pct_Parcels_w_Arts.Biz,
    rank
  )

x.industrial.spatial <- inner_join(
  mapc_municipalities,
  x.industrial,
  by = c("muni_name")
)

map.test <- ggplot() +
  geom_sf(data = x.industrial.spatial, aes(fill = pct_Parcels_w_Arts.Biz), color = "black")+
  scale_fill_continuous(high = "purple", low = "white", name = "% Industrial Parcels/nwith Arts Businesses") +
  labs(
    title = paste0("Percentage of Industrial Parcels with Arts Businesses (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.test

# Commercial

x.commercial <- x |> 
  filter(
    USE_CODE_SYMB %in% c("Commercial")
  ) |> 
  mutate(
    rank = (pct_Parcels_w_Arts.Biz - min(pct_Parcels_w_Arts.Biz))/ (max(pct_Parcels_w_Arts.Biz) - min(pct_Parcels_w_Arts.Biz))
  ) |> 
  select(
    muni_name,
    pct_Parcels_w_Arts.Biz,
    rank
  )

x.commercial.spatial <- inner_join(
  mapc_municipalities,
  x.commercial,
  by = c("muni_name")
)

map.test <- ggplot() +
  geom_sf(data = x.commercial.spatial, aes(fill = pct_Parcels_w_Arts.Biz), color = "black")+
  scale_fill_continuous(high = "purple", low = "white", name = "% Commercial Parcels/nwith Arts Businesses") +
  labs(
    title = paste0("Percentage of Commercial Parcels with Arts Businesses (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.test

# Residential

x.res <- x |> 
  filter(
    USE_CODE_SYMB %in% c("Residential")
  ) |> 
  mutate(
    rank = (pct_Parcels_w_Arts.Biz - min(pct_Parcels_w_Arts.Biz))/ (max(pct_Parcels_w_Arts.Biz) - min(pct_Parcels_w_Arts.Biz))
  ) |> 
  select(
    muni_name,
    pct_Parcels_w_Arts.Biz,
    rank
  )

x.res.spatial <- inner_join(
  mapc_municipalities,
  x.res,
  by = c("muni_name")
)

map.test <- ggplot() +
  geom_sf(data = x.res.spatial, aes(fill = pct_Parcels_w_Arts.Biz), color = "black")+
  scale_fill_continuous(high = "purple", low = "white", name = "% Residential Parcels/nwith Arts Businesses") +
  labs(
    title = paste0("Percentage of Residential Parcels with Arts Businesses (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.test
  
```


# 3.0 Tabular Analysis
# 3.1 Arts Businesses by Town and NAICS
```{r}
#Make a non-sf object copy of the spatialized arts business data to more
#easily facilitate tabular analysis.
data.axle_arts.spaces_mapc <- mapc_biz.spatial |> st_drop_geometry()

#Number of Arts Businesses by Municipality
arts.biz_by.town <- data.axle_arts.spaces_mapc |> 
  group_by(
    STCITY_CLEAN
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(businesses)
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.town,
  paste0(output_path, "arts.business_by_muni.csv")
)

#Create a table demonstrating the top 10 municipalities by number of arts businesses
tbl.1 <- arts.biz_by.town |>
  dplyr::slice(1:10) |> 
  gt() |> 
  gt::tab_header(
    title = "NEFA Arts Businesses by Municipality - MAPC"
  ) |> 
  gt::cols_label(
    STCITY_CLEAN = "Municipality",
    businesses = "Businesses",
    employment = "Employment",
    emp.by.biz = "Avg. Employment Per Business"
  ) |> 
  gt::fmt_number(
    columns = c("businesses", "employment"),
    sep_mark = ",",
    decimals = 0
  )

tbl.1 |> 
  gtsave(
    filename = paste0(output_path, "arts.biz_by.town.html")
  )

#Create a bar chart of Top 10 municipalities by number of businsesses
bar_graph.1 <- arts.biz_by.town |> 
  arrange(
    desc(businesses)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(STCITY_CLEAN, businesses), y = businesses)) +
  geom_bar(stat = "identity", color = "black", fill = "steelblue") +
  geom_text(aes(label = scales::comma(businesses)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Arts Businesses by Municipality (Top 10)",
    x = "Municipality",
    y = "Number of Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

#Check it in terminal
bar_graph.1

#Save
ggsave(
  filename = "top10_municipalities_by_arts.businesses.jpeg",
  plot = bar_graph.1,
  device = "jpeg",
  path = output_path
)

#Create a bar chart of Top 10 municipalities by employment
bar_graph.2 <- arts.biz_by.town |> 
  arrange(
    desc(employment)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(STCITY_CLEAN, employment), y = employment)) +
  geom_bar(stat = "identity", color = "black", fill = "lavender") +
  geom_text(aes(label = scales::comma(employment)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Arts Employment by Municipality (Top 10)",
    x = "Municipality",
    y = "Number of Employees"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.2

ggsave(
  filename = "top10_municipalities_by_arts.employment.jpeg",
  plot = bar_graph.2,
  device = "jpeg",
  path = output_path
)

#Create a bar chart of Top 10 municipalities by avg. employment per business
bar_graph.3 <- arts.biz_by.town |> 
  arrange(
    desc(emp.by.biz)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(STCITY_CLEAN, emp.by.biz), y = emp.by.biz)) +
  geom_bar(stat = "identity", color = "black", fill = "lightgreen") +
  geom_text(aes(label = scales::comma(emp.by.biz)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Avg. Arts Employment Per Businesses by Municipality (Top 10)",
    x = "Municipality",
    y = "Avg. Employment Per Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.3

ggsave(
  filename = "top10_municipalities_by_arts.employmentbybusiness.jpeg",
  plot = bar_graph.3,
  device = "jpeg",
  path = output_path
)

#Analysis
#99 of MAPC's 101 municipalities are home to at least one arts business.

#=======================================================================================
#Number of Arts Businesses by 6-Digit NAICS Code
arts.biz_by.naics <- data.axle_arts.spaces_mapc |> 
  group_by(
    NAICSD,
    NAICS.6
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(businesses)
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.naics,
  paste0(output_path, "arts.business_by_NAICS.csv")
)

#Create a table demonstrating the top 10 NAICS Codes by number of arts businesses
tbl.2 <- arts.biz_by.naics |>
  dplyr::slice(1:10) |> 
  gt() |> 
  tab_header(
    title = "NEFA Art Businesses by NAICS Code - MAPC"
  ) |> 
  cols_label(
    NAICSD = "NAICS",
    NAICS.6 = "6-Digit NAICS Code",
    businesses = "Businesses",
    employment = "Employment",
    emp.by.biz = "Avg. Employment Per Business"
  ) |> 
  fmt_number(
    columns = c("businesses", "employment"),
    sep_mark = ",",
    decimals = 0
  )

tbl.2 |> 
  gtsave(
    filename = paste0(output_path, "arts.biz_by.naics.html")
  )

#Create a bar chart of Top 10 municipalities by number of businsesses
bar_graph.1 <- arts.biz_by.naics |> 
  arrange(
    desc(businesses)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, businesses), y = businesses)) +
  geom_bar(stat = "identity", color = "black", fill = "steelblue") +
  geom_text(aes(label = scales::comma(businesses)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(
    limits = rev,
    labels = scales::label_wrap(10)
  ) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Arts Businesses by NAICS (Top 10)",
    x = "NAICS Description",
    y = "Number of Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

#Check it in terminal
bar_graph.1

#Save
ggsave(
  filename = "top10_naics_by_arts.businesses.jpeg",
  plot = bar_graph.1,
  device = "jpeg",
  path = output_path
)

#Create a bar chart of Top 10 municipalities by employment
bar_graph.2 <- arts.biz_by.naics |> 
  arrange(
    desc(employment)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, employment), y = employment)) +
  geom_bar(stat = "identity", color = "black", fill = "lavender") +
  geom_text(aes(label = scales::comma(employment)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Arts Employment by NAICS (Top 10)",
    x = "NAICS Description",
    y = "Number of Employees"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.2

ggsave(
  filename = "top10_naics_by_arts.employment.jpeg",
  plot = bar_graph.2,
  device = "jpeg",
  path = output_path
)

#Create a bar chart of Top 10 municipalities by avg. employment per business
bar_graph.3 <- arts.biz_by.naics |> 
  arrange(
    desc(emp.by.biz)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, emp.by.biz), y = emp.by.biz)) +
  geom_bar(stat = "identity", color = "black", fill = "lightgreen") +
  geom_text(aes(label = scales::comma(emp.by.biz)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(7)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Avg. Arts Employment Per Businesses by Municipality (Top 10)",
    x = "NAICS Description",
    y = "Avg. Employment Per Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.3

ggsave(
  filename = "top10_NAICS_by_arts.employmentbybusiness.jpeg",
  plot = bar_graph.3,
  device = "jpeg",
  path = output_path
)

#Analysis
#49 different types of arts business are represented in the MAPC region.


#==============================================================================
#Number of Arts businesses by municipality and 6-Digit NAICS Code
arts.biz_by.town.by.naics <- data.axle_arts.spaces_mapc |> 
  group_by(
    STCITY_CLEAN,
    NAICSD,
    NAICS.6
  ) |> 
  summarise(
    businesses = n(),
    employment = sum(LOCEMP),
    emp.by.biz = round(employment/businesses,2)
  ) |> 
  ungroup() |> 
  arrange(
    desc(desc(STCITY_CLEAN))
  )

#Write a csv of the above table
write.csv(
  arts.biz_by.town.by.naics,
  paste0(output_path, "arts.business_by_muni_by_NAICS.csv")
)

#Analysis
#
```

# 3.2 Arts Businesses by Land Use Codes
```{r}
#Density plot of employment of arts businesses in parcel
print(summary(biz_in_parcel |> select()))

biz_in_parcel |> 
  ggplot(aes(x = LOCEMP)) +
  geom_density() + 
  theme_minimal()

#Business by Minimum land use codes
Biz_by_LUC <- biz_in_parcel |> 
  group_by(
    USE_CODE_SYMB,
  ) |> 
  summarise(
    Num_Business = n(),
    Employment = sum(LOCEMP),
    Avg_emp = round(Employment/Num_Business, 2)
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry() |> 
  arrange(
    desc(Num_Business)
  )

write.csv(
  Biz_by_LUC,
  paste0(output_path, "businesses_by_LUC.type.csv")
)

#Create a bar chart of Top 10 municipalities by avg. employment per business
bar_graph.1 <- Biz_by_LUC |> 
  arrange(
    desc(Num_Business)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(USE_CODE_SYMB, Num_Business), y = Num_Business)) +
  geom_bar(stat = "identity", color = "black", fill = "steelblue") +
  geom_text(aes(label = scales::comma(Num_Business)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Businesses by Land Use Type (Top 10)",
    x = "Land Use Type",
    y = "Number of Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.1

ggsave(
  filename = "top10_LUC_by_arts.business.jpeg",
  plot = bar_graph.1,
  device = "jpeg",
  path = output_path
)

bar_graph.2 <- Biz_by_LUC |> 
  arrange(
    desc(Employment)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(USE_CODE_SYMB, Employment), y = Employment)) +
  geom_bar(stat = "identity", color = "black", fill = "lavender") +
  geom_text(aes(label = scales::comma(Employment)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Employment by Land Use Type (Top 10)",
    x = "Land Use Type",
    y = "Employment"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.2

ggsave(
  filename = "top10_LUC_by_arts.employment.jpeg",
  plot = bar_graph.2,
  device = "jpeg",
  path = output_path
)

bar_graph.3 <- Biz_by_LUC |> 
  arrange(
    desc(Avg_emp)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(USE_CODE_SYMB, Avg_emp), y = Avg_emp)) +
  geom_bar(stat = "identity", color = "black", fill = "lightgreen") +
  geom_text(aes(label = scales::comma(Avg_emp)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Average Employment per Business by Land Use Type (Top 10)",
    x = "Land Use Type",
    y = "Avg. Employment per Business"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.3

ggsave(
  filename = "top10_LUC_by_arts.avgempbybusiness.jpeg",
  plot = bar_graph.3,
  device = "jpeg",
  path = output_path
)

#Businesses by land use codes and 6-Digit NAICS
Biz_by_LUC_by_NAICS <- biz_in_parcel |> 
  group_by(
    USE_CODE_SYMB,
    NAICSD,
    NAICS.6
  ) |> 
  summarise(
    Num_Business = n(),
    Employment = sum(LOCEMP),
    Avg_emp = round(Employment/Num_Business, 2)
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry() |> 
  arrange(
    desc(Num_Business)
  )

write.csv(
  Biz_by_LUC_by_NAICS,
  paste0(output_path, "businesses_by_LUC.type_by_NAICS.csv")
)

Biz_on_Ind.Parcels_by_NAICS <- Biz_by_LUC_by_NAICS |> 
  filter(
    USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant"
  )

bar_graph.4 <- Biz_on_Ind.Parcels_by_NAICS |> 
  arrange(
    desc(Num_Business)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, Num_Business), y = Num_Business)) +
  geom_bar(stat = "identity", color = "black", fill = "steelblue") +
  geom_text(aes(label = scales::comma(Num_Business)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Number of Business on Industrial Parcels (Top 10, MAPC Region)",
    x = "NAICS Description",
    y = "Number of Businesses"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.4

ggsave(
  filename = "top10_Arts.Biz_Ind.Parcels.jpeg",
  plot = bar_graph.4,
  device = "jpeg",
  path = output_path
)

bar_graph.5 <- Biz_on_Ind.Parcels_by_NAICS |> 
  arrange(
    desc(Employment)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, Employment), y = Employment)) +
  geom_bar(stat = "identity", color = "black", fill = "lavender") +
  geom_text(aes(label = scales::comma(Employment)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Employment in Arts Businesses on Industrial Parcels (Top 10, MAPC Region)",
    x = "NAICS Description",
    y = "Employment"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.5

ggsave(
  filename = "top10_Arts.Emp_Ind.Parcels.jpeg",
  plot = bar_graph.5,
  device = "jpeg",
  path = output_path
)

bar_graph.6 <- Biz_on_Ind.Parcels_by_NAICS |> 
  arrange(
    desc(Avg_emp)
  ) |> 
  dplyr::slice(1:10) |> 
  ggplot(aes(x = reorder(NAICSD, Avg_emp), y = Avg_emp)) +
  geom_bar(stat = "identity", color = "black", fill = "lightgreen") +
  geom_text(aes(label = scales::comma(Avg_emp)), vjust = -0.3, size = 3.5) +
  scale_x_discrete(limits = rev, labels = scales::label_wrap(10)) +
  scale_y_continuous(label = scales::comma) +
  labs(
    title = "Avg. Employment in Arts Businesses on Industrial Parcels (Top 10, MAPC Region)",
    x = "NAICS Description",
    y = "Avg. Employment per Business"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(color = "black"),
    axis.text.y = element_text(color = "black")
  )

bar_graph.6

ggsave(
  filename = "top10_Arts.Avg.Emp.Biz_Ind.Parcels.jpeg",
  plot = bar_graph.6,
  device = "jpeg",
  path = output_path
)

#Businesses by Land Use Code and Municipality
Biz_by_LUC_Muni <- biz_in_parcel |> 
  group_by(
    USE_CODE_SYMB,
    muni_name
  ) |> 
  summarise(
    Num_Business = n(),
    Employment = sum(LOCEMP),
    Avg_emp = round(Employment/Num_Business, 2)
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry() |> 
  arrange(
    desc(Num_Business)
  )

write.csv(
 Biz_by_LUC_Muni,
  paste0(output_path, "businesses_by_LUC.type_by_Muni.csv")
)

#Businesses by Land Use Code, NAICS, and Municipality
Biz_by_LUC_by_NAICS_Muni <- biz_in_parcel |> 
  group_by(
    USE_CODE_SYMB,
    NAICSD,
    NAICS.6,
    muni_name
  ) |> 
  summarise(
    Num_Business = n(),
    Employment = sum(LOCEMP),
    Avg_emp = round(Employment/Num_Business, 2)
  ) |> 
  ungroup() |> 
  na.omit() |> 
  st_drop_geometry() |> 
  arrange(
    desc(Num_Business)
  )

write.csv(
 Biz_by_LUC_by_NAICS_Muni,
  paste0(output_path, "businesses_by_LUC.type_by_NAICS_by_Muni.csv")
)

```

# 4.0 Spatial Analysis
# 4.1 General Municipal Maps
```{r}
#Set the municipality that you would like to map
set_muni <- "Pembroke"

#Visualize the parcels that do have arts businesses in them
municipal_parcels <- mapc_parcels |> 
  filter(
    #Filter Parcels by the municipality set at the beginning of this chunk
    CITY == set_muni
  ) |> 
  #Join the count of arts businesses by parcel LOC_ID
  left_join(
    Parcel_Biz,
    by = c("LOC_ID")
  ) |> 
  mutate(
    #Set NA values to 0 for number of businesses and employment
    biz = tidyr::replace_na(biz, 0),
    emp = tidyr::replace_na(emp, 0),
    #Create bins for the number of businesses on a parcel for mapping later
    biz_bins = case_when(
      biz <= 0 ~ "0",
      biz == 1 ~ "1",
      biz == 2 ~"2",
      biz >= 3 & biz <= 5 ~ "3-5",
      biz >= 6 & biz <= 10 ~ "6-10",
      biz >= 11 & biz <= 15 ~ "11-15",
      biz >= 16 & biz <= 20 ~ "16-20",
      biz >= 21 & biz <= 25 ~ "21-25",
      biz >= 26 & biz <= 30 ~ "26-30",
      biz > 30 ~"> 30"
    ),
    #Binary representing parcels with businesses in industrial parcels
    ind_intersect = if_else(
      (USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant") & biz > 0, 1, 0
    )
  )

#Create a layer of solely industrial parcels
municipal_parcels_ind <- municipal_parcels |> 
  filter(
    USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant"
  )

#Filter for the municipal boundary of the selected municipality
muni_boundary <- mapc_municipalities |> 
  filter(
    muni_name == set_muni
  )

#Business point data
points <- mapc_biz.spatial |> 
  filter(
    STCITY_CLEAN == set_muni #& OBJECT_ID != "169672"
  )

#Map 1: Point locations of arts businsesses over a parcel map of a municipality
map.1 <- ggplot() +
  geom_sf(data = municipal_parcels, fill = NA, color = "grey")+
  geom_sf(data = points) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.1

ggsave(
  filename = paste0("BizPoints_",set_muni,".jpeg"),
  plot = map.1,
  device = "jpeg",
  path = output_path
)

#Map 2: Parcel arts business density map with overlay of industrial parcels
map.2 <- ggplot() +
  geom_sf(data = municipal_parcels, linewidth = 0.001, aes(fill = biz), color = "grey") +
  scale_fill_continuous(
    high = "purple", low = "white",
    breaks = c(
      min(waltham_parcels_full$biz),
      round(max(waltham_parcels_full$biz)*(1/3),0),
      round(max(waltham_parcels_full$biz)*(2/3),0),
      max(waltham_parcels_full$biz)),
    name = "Number of/nBusinesses"
  )+
  geom_sf(data = municipal_parcels_ind, aes(color = USE_CODE_SYMB), fill = NA, linewidth = 0.8)+
  scale_color_discrete(
    name = "Industrial Parcels"
  )+
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5)+
  labs(
    title = paste0("Density of Arts Businesses in ", set_muni, " (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.2

ggsave(
  filename = paste0("BizDensity_",set_muni,".jpeg"),
  plot = map.2,
  device = "jpeg",
  path = output_path
)

#Map 3: Highlights where arts businsesses intersect with industrial parcels
map.3 <- ggplot() +
  geom_sf(data = municipal_parcels, aes(fill = as.factor(ind_intersect)), color = "grey") +
  scale_fill_manual(
    values = c("white", "purple"),
    name = "Industrial Parcels/nwith Arts Businesses"
  ) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Intersect of Arts Businesses & Industrial Parcels in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.3

ggsave(
  filename = paste0("BizIndIntersect_",set_muni,".jpeg"),
  plot = map.3,
  device = "jpeg",
  path = output_path
)
```

4.2 Somerville Maps
```{r}
#Set the municipality that you would like to map
set_muni <- "Somerville"

#Visualize the parcels that do have arts businesses in them
municipal_parcels <- mapc_parcels |> 
  filter(
    #Filter Parcels by the municipality set at the beginning of this chunk
    CITY == set_muni
  ) |> 
  #Join the count of arts businesses by parcel LOC_ID
  left_join(
    Parcel_Biz,
    by = c("LOC_ID")
  ) |> 
  mutate(
    #Set NA values to 0 for number of businesses and employment
    biz = tidyr::replace_na(biz, 0),
    emp = tidyr::replace_na(emp, 0),
    #Create bins for the number of businesses on a parcel for mapping later
    biz_bins = case_when(
      biz <= 0 ~ "0",
      biz == 1 ~ "1",
      biz == 2 ~"2",
      biz >= 3 & biz <= 5 ~ "3-5",
      biz >= 6 & biz <= 10 ~ "6-10",
      biz >= 11 & biz <= 15 ~ "11-15",
      biz >= 16 & biz <= 20 ~ "16-20",
      biz >= 21 & biz <= 25 ~ "21-25",
      biz >= 26 & biz <= 30 ~ "26-30",
      biz > 30 ~"> 30"
    ),
    #Binary representing parcels with businesses in industrial parcels
    ind_intersect = if_else(
      (USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant") & biz > 0, 1, 0
    )
  )

#Create a layer of solely industrial parcels
municipal_parcels_ind <- municipal_parcels |> 
  filter(
    USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant"
  )

#Filter for the municipal boundary of the selected municipality
muni_boundary <- mapc_municipalities |> 
  filter(
    muni_name == set_muni
  )

#Business point data
points <- mapc_biz.spatial |> 
  filter(
    STCITY_CLEAN == set_muni #& OBJECT_ID != "169672"
  )

#Map 1: Point locations of arts businsesses over a parcel map of a municipality
map.1 <- ggplot() +
  geom_sf(data = municipal_parcels, fill = NA, color = "grey")+
  geom_sf(data = points) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.1

ggsave(
  filename = paste0("BizPoints_",set_muni,".jpeg"),
  plot = map.1,
  device = "jpeg",
  path = output_path
)

#Map 2: Parcel arts business density map with overlay of industrial parcels
map.2 <- ggplot() +
  geom_sf(data = municipal_parcels, linewidth = 0.001, aes(fill = biz), color = "grey")+
  scale_fill_continuous(
    high = "purple", low = "white",
    breaks = c(
      min(municipal_parcels$biz),
      round(max(municipal_parcels$biz)*(1/3),0),
      round(max(municipal_parcels$biz)*(2/3),0),
      max(municipal_parcels$biz)),
    name = "Number of/nBusinesses"
  )+
  geom_sf(data = municipal_parcels_ind, aes(color = USE_CODE_SYMB), fill = NA, linewidth = 0.8)+
  scale_color_discrete(
    name = "Industrial Parcels"
  )+
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5)+
  labs(
    title = paste0("Density of Arts Businesses in ", set_muni, " (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.2

ggsave(
  filename = paste0("BizDensity_",set_muni,".jpeg"),
  plot = map.2,
  device = "jpeg",
  path = output_path
)

#Map 3: Highlights where arts businsesses intersect with industrial parcels
map.3 <- ggplot() +
  geom_sf(data = municipal_parcels, aes(fill = as.factor(ind_intersect)), color = "grey") +
  scale_fill_manual(
    values = c("white", "purple"),
    name = "Industrial Parcels/nwith Arts Businesses"
  ) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Intersect of Arts Businesses & Industrial Parcels in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.3

ggsave(
  filename = paste0("BizIndIntersect_",set_muni,".jpeg"),
  plot = map.3,
  device = "jpeg",
  path = output_path
)

map.4 <- ggplot() +
  geom_sf(data = municipal_parcels, fill = NA, color = "grey")+
  geom_sf(data = points) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  geom_sf(data = fab_dist, fill = NA, color = "maroon", linewidth = 0.8) +
  scale_color_discrete(
    name = "Fabrication Zoning District"
  ) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", set_muni, " (2023)"),
    subtitle = "Fabrication Zoning District Boundaries (maroon)"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.4

ggsave(
  filename = paste0("BizPoint_FAB.District_",set_muni,".jpeg"),
  plot = map.4,
  device = "jpeg",
  path = output_path
)
```

# 4.3 Cambridge Maps
```{r}
#Set the municipality that you would like to map
set_muni <- "Cambridge"

#Visualize the parcels that do have arts businesses in them
municipal_parcels <- mapc_parcels |> 
  filter(
    #Filter Parcels by the municipality set at the beginning of this chunk
    CITY == set_muni
  ) |> 
  #Join the count of arts businesses by parcel LOC_ID
  left_join(
    Parcel_Biz,
    by = c("LOC_ID")
  ) |> 
  mutate(
    #Set NA values to 0 for number of businesses and employment
    biz = tidyr::replace_na(biz, 0),
    emp = tidyr::replace_na(emp, 0),
    #Create bins for the number of businesses on a parcel for mapping later
    biz_bins = case_when(
      biz <= 0 ~ "0",
      biz == 1 ~ "1",
      biz == 2 ~"2",
      biz >= 3 & biz <= 5 ~ "3-5",
      biz >= 6 & biz <= 10 ~ "6-10",
      biz >= 11 & biz <= 15 ~ "11-15",
      biz >= 16 & biz <= 20 ~ "16-20",
      biz >= 21 & biz <= 25 ~ "21-25",
      biz >= 26 & biz <= 30 ~ "26-30",
      biz > 30 ~"> 30"
    ),
    #Binary representing parcels with businesses in industrial parcels
    ind_intersect = if_else(
      (USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant") & biz > 0, 1, 0
    )
  )

#Create a layer of solely industrial parcels
municipal_parcels_ind <- municipal_parcels |> 
  filter(
    USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant"
  )

#Filter for the municipal boundary of the selected municipality
muni_boundary <- mapc_municipalities |> 
  filter(
    muni_name == set_muni
  )

#Business point data
points <- mapc_biz.spatial |> 
  filter(
    STCITY_CLEAN == set_muni #& OBJECT_ID != "169672"
  )

#Map 1: Point locations of arts businsesses over a parcel map of a municipality
map.1 <- ggplot() +
  geom_sf(data = municipal_parcels, fill = NA, color = "grey")+
  geom_sf(data = points) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.1

ggsave(
  filename = paste0("BizPoints_",set_muni,".jpeg"),
  plot = map.1,
  device = "jpeg",
  path = output_path
)

#Map 2: Parcel arts business density map with overlay of industrial parcels
map.2 <- ggplot() +
  geom_sf(data = municipal_parcels, linewidth = 0.001, aes(fill = biz), color = "grey")+
  scale_fill_continuous(
    high = "purple", low = "white",
    breaks = c(
      min(municipal_parcels$biz),
      round(max(municipal_parcels$biz)*(1/3),0),
      round(max(municipal_parcels$biz)*(2/3),0),
      max(municipal_parcels$biz)),
    name = "Number of Businesses"
  )+
  geom_sf(data = municipal_parcels_ind, aes(color = USE_CODE_SYMB), fill = NA, linewidth = 0.4)+
  scale_color_discrete(
    name = "Industrial Parcels"
  )+
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5)+
  labs(
    title = paste0("Density of Arts Businesses in ", set_muni, " (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.2

ggsave(
  filename = paste0("BizDensity_",set_muni,".jpeg"),
  plot = map.2,
  device = "jpeg",
  path = output_path,
  height = 5,
  width = 10,
  dpi = 300
)

#Map 3: Highlights where arts businsesses intersect with industrial parcels
map.3 <- ggplot() +
  geom_sf(data = municipal_parcels, aes(fill = as.factor(ind_intersect)), color = "grey") +
  scale_fill_manual(
    values = c("white", "purple"),
    name = "Industrial Parcels/nwith Arts Businesses"
  ) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Intersect of Arts Businesses & Industrial Parcels in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.3

ggsave(
  filename = paste0("BizIndIntersect_",set_muni,".jpeg"),
  plot = map.3,
  device = "jpeg",
  path = output_path
)

#Map 4: Point locations of arts businsesses over a parcel map of a municipality with neighborhoods
muni_nbhd <- nbhds |> 
  filter(
    muni == set_muni & Nhood == "North Cambridge"
  )

nhood_parcels <- municipal_parcels[muni_nbhd,]

nhood_points <- points[muni_nbhd,]

muni_nbhd_centroids <- nbhds_centroids |> 
  filter(
    muni == set_muni
  )

#East Cambridge
zoom_to <- muni_nbhd_centroids |>
  filter(Nhood == "North Cambridge") |>
  extract(geometry, into = c('Lat', 'Lon'), '\\((.*),(.*)\\)', conv = T) |>
  select(
    Lat,
    Lon
  ) |> 
  mutate(
    across(everything(), as.numeric)
  )

zoom_level <- 7

lon_span <- 360/2^zoom_level
lat_span <- 180/2^zoom_level

lon_bounds <- c(as.numeric(zoom_to[1] - lon_span/2)[1], as.numeric(zoom_to[1] + lon_span/2)[1])
lat_bounds <- c(as.numeric(zoom_to[2] - lat_span/2)[1], as.numeric(zoom_to[2] + lat_span/2)[1])

lon_min <- as.numeric(zoom_to[1] - lon_span/2)[1]
lon_max <- as.numeric(zoom_to[1] + lon_span/2)[1]
lat_min <- as.numeric(zoom_to[2] - lat_span/2)[1]
lat_max <- as.numeric(zoom_to[2] + lat_span/2)[1]

municipal_parcels_cropped <- st_crop(
  municipal_parcels,
  xmin = lon_min,
  xmax = lon_max,
  ymin = lat_min,
  ymax = lat_max
)

zoom_to <- zoom_to |> 
  #Converting the non-spatial data into a spatial object.
  st_as_sf(
    #coords = c("LONGITUDE", "LATITUDE"),
    coords = c("Lon","Lat"),
    crs = st_crs(lat_lon_CRS),
    remove = FALSE
  ) |>
  #Setting CRS to match the polygon's CRS.
  st_transform(
    mass_mainland,
    crs = st_crs(mapc_parcels)
  )

#view(lon_bounds)

map.4 <- ggplot() +
  geom_sf(data = nhood_parcels, fill = NA, color = "grey")+
  geom_sf(data = nhood_points) +
  #geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  geom_sf(data = muni_nbhd, fill = NA, color = "black", linewidth = 0.5) +
  #geom_sf(data = municipal_parcels_cropped) +
  #geom_sf(data = zoom_to) +
  #coord_sf(xlim = lon_bounds, ylim = lat_bounds) +
  #geom_label_repel(
    #data = muni_nbhd,
    #aes(label = Nhood, geometry = geometry),
    #stat = "sf_coordinates",
    #min.segment.length = 0
  #) +
  labs(
    title = paste0("Point Location of Arts Businesses in North Cambridge (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.4

ggsave(
  filename = paste0("BizPoints_Nbhd",set_muni,".jpeg"),
  plot = map.4,
  device = "jpeg",
  path = output_path
)
```

# 4.4 Boston Maps
```{r}
#Set the municipality that you would like to map
set_muni <- "Boston"

#Visualize the parcels that do have arts businesses in them
municipal_parcels <- mapc_parcels |> 
  filter(
    #Filter Parcels by the municipality set at the beginning of this chunk
    CITY == set_muni
  ) |> 
  #Join the count of arts businesses by parcel LOC_ID
  left_join(
    Parcel_Biz,
    by = c("LOC_ID")
  ) |> 
  mutate(
    #Set NA values to 0 for number of businesses and employment
    biz = tidyr::replace_na(biz, 0),
    emp = tidyr::replace_na(emp, 0),
    #Create bins for the number of businesses on a parcel for mapping later
    biz_bins = case_when(
      biz <= 0 ~ "0",
      biz == 1 ~ "1",
      biz == 2 ~"2",
      biz >= 3 & biz <= 5 ~ "3-5",
      biz >= 6 & biz <= 10 ~ "6-10",
      biz >= 11 & biz <= 15 ~ "11-15",
      biz >= 16 & biz <= 20 ~ "16-20",
      biz >= 21 & biz <= 25 ~ "21-25",
      biz >= 26 & biz <= 30 ~ "26-30",
      biz > 30 ~"> 30"
    ),
    #Binary representing parcels with businesses in industrial parcels
    ind_intersect = if_else(
      (USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant") & biz > 0, 1, 0
    )
  )

#Create a layer of solely industrial parcels
municipal_parcels_ind <- municipal_parcels |> 
  filter(
    USE_CODE_SYMB == "Industrial" | USE_CODE_SYMB == "Industrial Vacant"
  )

#Filter for the municipal boundary of the selected municipality
muni_boundary <- mapc_municipalities |> 
  filter(
    muni_name == set_muni
  )

#Business point data
points <- mapc_biz.spatial |> 
  filter(
    STCITY_CLEAN == set_muni #& OBJECT_ID != "169672"
  )

#Map 1: Point locations of arts businsesses over a parcel map of a municipality
map.1 <- ggplot() +
  geom_sf(data = municipal_parcels, fill = NA, color = "grey")+
  geom_sf(data = points) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.1

ggsave(
  filename = paste0("BizPoints_",set_muni,".jpeg"),
  plot = map.1,
  device = "jpeg",
  path = output_path
)

#Map 2: Parcel arts business density map with overlay of industrial parcels
map.2 <- ggplot() +
  geom_sf(data = municipal_parcels, linewidth = 0.001, aes(fill = biz), color = "grey")+
  scale_fill_continuous(
    high = "purple", low = "white",
    breaks = c(
      min(waltham_parcels_full$biz),
      round(max(waltham_parcels_full$biz)*(1/3),0),
      round(max(waltham_parcels_full$biz)*(2/3),0),
      max(waltham_parcels_full$biz)),
    name = "Number of/nBusinesses"
  )+
  geom_sf(data = municipal_parcels_ind, aes(color = USE_CODE_SYMB), fill = NA, linewidth = 0.8)+
  scale_color_discrete(
    name = "Industrial Parcels"
  )+
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5)+
  labs(
    title = paste0("Density of Arts Businesses in ", set_muni, " (2023)")
  )+
  theme_minimal()+
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.2

ggsave(
  filename = paste0("BizDensity_",set_muni,".jpeg"),
  plot = map.2,
  device = "jpeg",
  path = output_path
)

#Map 3: Highlights where arts businsesses intersect with industrial parcels
map.3 <- ggplot() +
  geom_sf(data = municipal_parcels, aes(fill = as.factor(ind_intersect)), color = "grey") +
  scale_fill_manual(
    values = c("white", "purple"),
    name = "Industrial Parcels/nwith Arts Businesses"
  ) +
  geom_sf(data = muni_boundary, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Intersect of Arts Businesses & Industrial Parcels in ", set_muni, " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.3

ggsave(
  filename = paste0("BizIndIntersect_",set_muni,".jpeg"),
  plot = map.3,
  device = "jpeg",
  path = output_path
)

#Map 4:

muni_nbhd <- nbhds |> 
  filter(
    muni == set_muni & Nhood == "Allston/Brighton"
  )

nhood_parcels <- municipal_parcels[muni_nbhd,]

nhood_points <- points[muni_nbhd,]

map.4 <- ggplot() +
  geom_sf(data = nhood_parcels, fill = NA, color = "grey")+
  geom_sf(data = nhood_points) +
  geom_sf(data = muni_nbhd, fill = NA, color = "black", linewidth = 0.5) +
  labs(
    title = paste0("Point Location of Arts Businesses in ", muni_nbhd[2], " (2023)")
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_blank(),
    axis.text.y = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(fill = NA, color = "black", linewidth = 0.5)
  )

map.4
```
# 5.0 Statitsical Testing
# 5.1 Test for Spatial Dependence - Does Clustering Happen?
```{r}
#We can visually inspect the maps we've made to judge whether or not arts businsses
#co-locate. We can test this assumption by determining the spatial dependence
#of arts businesses at the parcel level.

#First we need to determine which polygons (parcels) are neighbors to one another.
#There are two common methods for this: rook and queen contiguity. Rook contiguity refers
#to parcels with borders in contact with each other. Queen contiguity refers to parcels with borders that are near to each other in all directions.

#===================================================================================
#TODO:
# 1. Create loop which tests all MAPC municipalities
# 2. Save results of test to a single data.frame
# 3. test MAPC as a whole
#===================================================================================
# MAPC-wide Test
#Create a separate set of parcel data for the test municipality
#Join the count of arts businesses by parcel LOC_ID
cluster_anlaysis_parcels <- left_join(
    mapc_parcels,
    Parcel_Biz,
    by = c("LOC_ID")
  ) |> 
  mutate(
    #Set NA values to 0 for number of businesses and employment
    biz = tidyr::replace_na(biz, 0),
    emp = tidyr::replace_na(emp, 0),
  )

rook_nhbrs <- spdep::poly2nb(cluster_anlaysis_parcels, queen = FALSE)
  
queen_nhbrs <- spdep::poly2nb(cluster_anlaysis_parcels, queen = TRUE)

#Second we need to create a spatial weight matrix to quantify the location of the
#parcel polygons. Parcels close to one another will have weights closer to 1; those
#further away will have weight values closer to 0.
lw_rook <- nb2listw(rook_nhbrs, style = "W", zero.policy = TRUE)

lw_queen <- nb2listw(queen_nhbrs, style = "W", zero.policy = TRUE)

#Sidebar

biz.lag_rook <- lag.listw(lw_rook, cluster_anlaysis_parcels$biz)

plot(biz.lag_rook ~ cluster_anlaysis_parcels$biz, pch = 16, asp = 1)
line.1 <- lm(biz.lag_rook ~ cluster_anlaysis_parcels$biz)
abline(line.1, col = "blue")

biz.lag_queen <- lag.listw(lw_queen, cluster_anlaysis_parcels$biz)

plot(biz.lag_queen ~ cluster_anlaysis_parcels$biz, pch = 16, asp = 1)
line.2 <- lm(biz.lag_queen ~ cluster_anlaysis_parcels$biz)
abline(line.1, col = "blue")

#

#Third we calculate a Moran's I statistic for each parcel based on the density of 
#arts business in each parcel. Fundamentally this is a weighted covariance. The weights
#are those calculated in the spatial weight matrix. Moran's I falls betwee -1 and 1. Values closer to either #minimum or maximum indicate covariance -- the spatial objects are alike or the value 
#being considered is a function of that value on the other spatial object. #Values closer to 0
#indicate two spatial objects are not alike.
I_rook <- moran(cluster_anlaysis_parcels$biz, lw_rook, length(rook_nhbrs), Szero(lw_rook))
I_rook[1]

I_queen <- moran(cluster_anlaysis_parcels$biz, lw_queen, length(queen_nhbrs), Szero(lw_queen))
I_queen[1]

#Fourth we conduct a hypothesis test. It works like a one sided t-test.
moran.test(cluster_anlaysis_parcels$biz, lw_rook, alternative = "greater")

moran.test(cluster_anlaysis_parcels$biz, lw_queen, alternative = "greater")

#===================================================================================
#Individual Municipality Test

munis <- mapcdatakeys::all_muni_data_keys |> filter(mapc == 1 & !muni_name %in% c("Hamilton", "Manchester-by-the-Sea")) |>  select(muni_name) |> pull()

I_rook.list <- NULL
Moran.test_rook.list <- NULL

I_queen.list <- NULL
Moran.test_queen.list <- NULL
  
for (i in munis) {
  
  print(paste0("Testing Spatial Autocorrelation for ", i,"."))
  #Create a separate set of parcel data for the test municipality
  cluster_anlaysis_parcels <- mapc_parcels |>
    #Filter Parcels by the municipality set at the beginning of this chunk
    filter(CITY == i) |>
    #Join the count of arts businesses by parcel LOC_ID
    left_join(Parcel_Biz, by = c("LOC_ID")) |>
    #Set NA values to 0 for number of businesses and employment
    mutate(
      biz = tidyr::replace_na(biz, 0),
      emp = tidyr::replace_na(emp, 0)
    )
  # First, determine the neighbor matrix for each parcel using the rook and queen method.
  rook_nhbrs <- spdep::poly2nb(cluster_anlaysis_parcels, queen = FALSE)
  
  queen_nhbrs <- spdep::poly2nb(cluster_anlaysis_parcels, queen = TRUE)
  
  # Second we need to create a spatial weight matrix to quantify the location of the
  # parcel polygons. Parcels close to one another will have weights closer to 1; those
  # further away will have weight values closer to 0.
  lw_rook <- nb2listw(rook_nhbrs, style = "W", zero.policy = TRUE)
  
  lw_queen <- nb2listw(queen_nhbrs, style = "W", zero.policy = TRUE)
  
  #Sidebar
  
  # biz.lag_rook <- lag.listw(lw_rook, cluster_anlaysis_parcels$biz)
  # 
  # plot(biz.lag_rook ~ cluster_anlaysis_parcels$biz, pch = 16, asp = 1)
  # line.1 <- lm(biz.lag_rook ~ cluster_anlaysis_parcels$biz)
  # abline(line.1, col = "blue")
  # 
  # biz.lag_queen <- lag.listw(lw_queen, cluster_anlaysis_parcels$biz)
  # 
  # plot(biz.lag_queen ~ cluster_anlaysis_parcels$biz, pch = 16, asp = 1)
  # line.2 <- lm(biz.lag_queen ~ cluster_anlaysis_parcels$biz)
  # abline(line.1, col = "blue")
  
  #Third we calculate a Moran's I statistic for each parcel based on the density of
  #arts business in each parcel. Fundamentally this is a weighted covariance. The weights
  #are those calculated in the spatial weight matrix. Moran's I falls betwee -1 and 1. Values closer to either #minimum or maximum indicate covariance -- the spatial objects are alike or the value
  #being considered is a function of that value on the other spatial object. #Values closer to 0
  #indicate two spatial objects are not alike.
  # 1. Rook Moran's I stat generation
  I_rook <- moran(cluster_anlaysis_parcels$biz, lw_rook, length(rook_nhbrs), Szero(lw_rook))
  print(I_rook[1])
  
  # Collects the Moran I stat for the municipality being considered into a vector which will be aggregated later for review.
  I_rook.list <- bind_rows(I_rook.list, I_rook[1])
  
  # 2. Queen Moran's I stat generation
  I_queen <- moran(cluster_anlaysis_parcels$biz, lw_queen, length(queen_nhbrs), Szero(lw_queen))
  print(I_queen[1])
  
    # Collects the Moran I stat for the municipality being considered into a vector which will be aggregated later for review.
   I_queen.list <- bind_rows(I_queen.list, I_queen[1])
  
  # Fourth we conduct a hypothesis test. It works like a one sided t-test.
  # 1. Rook method hypothesis test
  mt.r <- moran.test(cluster_anlaysis_parcels$biz, lw_rook, alternative = "greater")
  
  # Print the summary of the test.
  print(mt.r)
  
    # Collects the p value for the Moran I stat hypothesis test for the municipality being considered into a vector which will be aggregated later for review. 
  Moran.test_rook.list <- bind_rows(Moran.test_rook.list, mt.r[2])
  
  # 2. Queen method hypothesis test
  mt.q <- moran.test(cluster_anlaysis_parcels$biz, lw_queen, alternative = "greater")
  
  # Print the summary of the test.
  print(mt.q)
  
    # Collects the p value for the Moran I stat hypothesis test for the municipality being considered into a vector which will be aggregated later for review. 
  Moran.test_queen.list <- bind_rows(Moran.test_queen.list, mt.q[2])
  
}

# Create a dataframe of muni names to bind to the associated stat vectors from the Moran's I tests in the for loop above.
muni_names <- data.frame(munis)

# Bind columns of muni names, Moran's I statistics, and p values from the Moran Tests for each municipality created in the for loop above.
test.stat.df <- bind_cols(
  muni_names,
  I_rook.list, Moran.test_rook.list,
  I_queen.list, Moran.test_queen.list
) |> 
  # Rename columns for comprehension.
  dplyr::rename(
    Rook_MoranI = I...2,
    Rook_MoranTest_pval = p.value...3,
    Queen_MoranI = I...4,
    Queen_MoranTest_pval = p.value...5,
  )

# General gist -
# MAPC as a region - yes businesses tend to cluster in parcels but those clusters are dispersed throughout the region, there isnt one area that's a central locus for arts businesses.
# Individual municipalities - It depends! Some do not exhibit clustering phenomenon, some do! Some places the clusters are concentrated - I stat closer to 1, others more dispersed - i stat closer to -1. And some places where the businesses tend to be evenly spread throughout the parcels in the municipality.

```
# 5.2 Classification of Clusters
```{r}
#===================================================================================
# TODO:
# 1. Figure out how to use distance as the factor to create KNN groups
#===================================================================================
```
